+++ efi_dxe_emulator/mem_taint.h
@@ -4,9 +4,9 @@
 
 
 bool taint_mem(uint64_t address);
-bool taint_mem(uint64_t address, uint8_t size);
+bool taint_mem(uint64_t address, uint32_t size);
 
-bool untaint_mem(uint64_t address, uint8_t size);
+bool untaint_mem(uint64_t address, uint32_t size);
 
 bool is_mem_tainted(uint64_t address);
-bool is_mem_tainted(uint64_t address, uint8_t size);
+bool is_mem_tainted(uint64_t address, uint32_t size);
